+++ service/database/delete-comment.go
@@ -1,6 +1,6 @@
 package database
 
-func (db *appdbimpl) DeleteComment(commentId int) error {
+func (db *appdbimpl) DeleteComment(commentId uint64) error {
 	_, err := db.c.Exec("DELETE FROM Comments WHERE Id = ?", commentId)
 	return err
 }
