+++ service/database/follow-user.go
@@ -1,7 +1,7 @@
 package database
 
 // GetName is an example that shows you how to query data
-func (db *appdbimpl) FollowUser(IdUserToFollow int, IdFollowingUser int) error {
+func (db *appdbimpl) FollowUser(IdUserToFollow uint64, IdFollowingUser uint64) error {
 	_, err := db.c.Exec(
 		"INSERT INTO Follows(FollowerId, FollowedId) VALUES (?, ?)", IdFollowingUser, IdUserToFollow)
 	return err
